Avoid sending RTP stream through sed
The RTP stream doesn't necessarily contain newline characters, so if we
send the RTP stream through sed, it just buffers a bunch of the camera
stream without writing anything for a while.

So once we detect that the camera is sending us the RTP stream, we just
send it through to the client without any processing.